replaced ints with uint16_t stuff

+++ src/cursor.cpp
@@ -1,7 +1,9 @@
 #include <SDL2/SDL.h>
+#include <stdint.h>
+
 #include "cursor.hpp"
 
-void CCursor::changeCursor(int x, int y, int w, int h)
+void CCursor::changeCursor(int x, int y, uint8_t w, uint8_t h)
 {
     cursor.x = (x-w)/scale;
     cursor.y = (y-h)/scale;
